pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inventory-ai-sota"
version = "1.0.0"
description = "A High-Fidelity Supply Chain Environment for the Meta OpenEnv Hackathon"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "numpy",
    "pydantic",
    "google-generativeai",
    "fastapi",
    "uvicorn",
    "httpx",
    "openai",
    "python-dotenv"
]

[project.scripts]
server = "server.app:main"


[tool.setuptools]
packages = ["server"]

